% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{as.list,dsoParams-method}
\alias{as.list,dsoParams-method}
\title{Custom as.list method for dsoParams class}
\usage{
\S4method{as.list}{dsoParams}(x)
}
\arguments{
\item{x}{dsoParams object}
}
\value{
A list
}
\description{
Custom as.list method for dsoParams class
}
