% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{$.dsoParams}
\alias{$.dsoParams}
\title{overriding the the $ operator to add secure list calling so that
it cannot return NULL when call does not exist}
\usage{
\method{$}{dsoParams}(x, name)
}
\arguments{
\item{x}{dsoParams object}

\item{name}{field name}
}
\description{
overriding the the $ operator to add secure list calling so that
it cannot return NULL when call does not exist
}
