% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_config.R
\name{compile_config}
\alias{compile_config}
\title{compile_config}
\usage{
compile_config(dir = getwd())
}
\description{
Run dso compile-config
}
\details{
This function runs the dso compile-config command and updates the params.yaml with info from other params.in.yaml and params.yaml.
}
\examples{
\dontrun{
compile_config()
}

}
\keyword{compile}
\keyword{config}
\keyword{dso}
\keyword{dvc}
\keyword{yaml}
