% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stage}
\alias{create_stage}
\title{create_stage}
\usage{
create_stage(dir = here::here(), name = "", description = "")
}
\arguments{
\item{dir}{path to directory in which stage shall be created}

\item{name}{stage name: e.g. 01_preprocessing. Can't be empty}

\item{description}{short stage description. Can't be empty}
}
\description{
creates new stage
}
\details{
Creates a new stage in a given directory.
}
\examples{
\dontrun{
create_stage(name = "01_preprocessing", description = "Sequencing Preprocessing")
}

}
\keyword{dso}
\keyword{dvc}
\keyword{parameter}
\keyword{yaml}
