% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{.apply_watermark}
\alias{.apply_watermark}
\title{Apply watermark to an image file using the dso CLI}
\usage{
.apply_watermark(input_file, output_file, watermark_config)
}
\arguments{
\item{input_file}{Path to the input image}

\item{output_file}{Path to the output (watermarked) image}

\item{watermark_config}{List with watermark configuration}
}
\description{
Apply watermark to an image file using the dso CLI
}
\keyword{internal}
