% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_get.R
\name{.convert_list_call_to_dollar_format}
\alias{.convert_list_call_to_dollar_format}
\title{safe_get}
\usage{
.convert_list_call_to_dollar_format(input_string, env = caller_env())
}
\arguments{
\item{input_string}{input string from deparsed function call}

\item{env}{environment which is used to resolve variables in the list call}
}
\value{
converted nested list call in $ format
}
\description{
converts a nested list call to dollar format
}
\details{
Converts a nested list call to dollar format while resolving variables
in the environment specified in `env` (by default the caller_env())

input_string <- 'CONFIG[[a]][[\"foo\"]]$bar[[\'no\']][[b]]$level'
a <- "test"
b <- "bla"
print(.convert_list_call_to_dollar_format('CONFIG[[a]][[\"foo\"]]$bar[["no"]][["further"]]$level'))
}
