% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{.get_watermark_config}
\alias{.get_watermark_config}
\title{Get watermark configuration from the global config environment}
\usage{
.get_watermark_config()
}
\value{
A list with watermark configuration or NULL
}
\description{
Get watermark configuration from the global config environment
}
\keyword{internal}
