% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{dso_repro_addin}
\alias{dso_repro_addin}
\title{Execute dso repro and display the result}
\usage{
dso_repro_addin()
}
\value{
None
}
\description{
This function runs the `dso repro` command on the specified `dvc.yaml` file
and displays the resulting report in the RStudio Viewer pane.
}
\examples{
\dontrun{
dso_repro_addin()
}
}
