% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{dso_repro_stage_w_dependencies_addin}
\alias{dso_repro_stage_w_dependencies_addin}
\title{Execute dso repro with dependencies and display the result}
\usage{
dso_repro_stage_w_dependencies_addin()
}
\value{
None
}
\description{
This function reproduces the current stage along its dependencies
and displays the resulting report in the RStudio Viewer pane.
}
\examples{
\dontrun{
dso_repro_stage_w_dependencies_addin()
}
}
