% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\docType{class}
\name{dsoParams-class}
\alias{dsoParams-class}
\title{dsoParams class}
\description{
dsoParams class
}
\section{Slots}{

\describe{
\item{\code{params}}{A list of lists containing parameters}
}}

