% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{dsoParams}
\alias{dsoParams}
\title{dsoParams: list with safe access}
\usage{
dsoParams(x = list())
}
\arguments{
\item{x}{empty, or a recursive list of lists which is converted to dsoParams}
}
\description{
A "dsoParams" and its constructor:
}
\examples{
# initiating empty
params <- dsoParams()

# converting a list of list

params <- list()
params$a <- "bla"
params$b <- list()
params$b$c <- "blub"
params <- dsoParams(params)

}
