% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{init}
\usage{
init(dir = here::here(), name = "", description = "")
}
\arguments{
\item{dir}{path to directory in which project shall be initialised}

\item{name}{project name: e.g. single_cell_lung_atlas. Can't be empty}

\item{description}{short project description. Can't be empty}
}
\description{
initialise new project
}
\details{
Initialises a new project in a given directory.
}
\examples{
\dontrun{
init(name = "single_cell_lung_atlas", description = "This analysis solves all our problems!")
}

}
\keyword{dso}
\keyword{dvc}
\keyword{parameter}
\keyword{project}
\keyword{yaml}
