% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_params}
\alias{read_params}
\title{read_params}
\usage{
read_params(stage_path, return_list = FALSE, quiet = FALSE)
}
\arguments{
\item{stage_path}{relative path to stage directory from project root}

\item{return_list}{returns a list if TRUE, by default it return `dsoParams` class which is a list with secure access}

\item{quiet}{suppresses the warnings and errors of the `dso get-config` execution}
}
\value{
parameters as list of list as `dsoParams` or conventional list when `return_list` is set.
}
\description{
Set stage dir and load parameters from params.yaml
}
\details{
It is required to provide the path of the current stage relative to the project root to ensure that
the correct config is loaded, no matter of the current working directory (as long as the working directory
is any subdirectory of the project root). The function recompiles params.in.yaml to params.yaml on-the-fly
to ensure that up-to-date params are always loaded.
}
