% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_safe_yaml}
\alias{read_safe_yaml}
\title{read_safe_yaml}
\usage{
read_safe_yaml(params_file)
}
\arguments{
\item{params_file}{path to yaml file}
}
\value{
a list
}
\description{
Read in YAML files while not interpolating yes/no and Y/N.
This is necessary because there's no YAML 1.2-compliant parser in R yet.
}
