% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{reload.dsoParams}
\alias{reload.dsoParams}
\title{Create a new method "reload" for class dsoParams}
\usage{
\method{reload}{dsoParams}(object)
}
\arguments{
\item{object}{dsoParams object}
}
\description{
Create a new method "reload" for class dsoParams
}
