% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro.R
\name{repro}
\alias{repro}
\title{Reproduce a dso stage}
\usage{
repro(stage_path = stage_here(), single_stage = F)
}
\arguments{
\item{stage_path}{The path to a stage. Defaults to the current stage by using `stage_here()`.}

\item{single_stage}{Logical flag indicating whether to reproduce only the current stage (`TRUE`) or the current stage with all dependencies (`FALSE`). Defaults to `FALSE`.}
}
\description{
Reproduce a dso stage with or without its dependencies.
}
\details{
This function reproduces a stage specified by `stage_path`. If `single_stage` is set to `TRUE`,
it reproduces the stage without its dependencies. Otherwise, it reproduces the current stage along with all its dependency stages.
By default, the current stage will be reproduced.
}
