% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{set_stage}
\alias{set_stage}
\title{set_stage}
\usage{
set_stage(stage_path)
}
\arguments{
\item{stage_path}{relative path to stage directory from project root}
}
\value{
absolute path to stage
}
\description{
Set stage dir
}
\details{
It is required to provide the path of the current stage relative to the project root. The stage
dir is set, which allows the use of stage_here and the config_env variable stage_dir.
To get the path to the stage use stage_here(). stage_here() will return the absolute path to the
stage from the project root.
}
