% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{stage_here}
\alias{stage_here}
\title{stage_here}
\usage{
stage_here(...)
}
\arguments{
\item{...}{additional parts of the path appended to the stage path using `file.path`}
}
\value{
absolute path to stage
}
\description{
get absolute path to stage
}
\details{
Get the absolute path to the current stage. The current stage can be set using set_stage()
by providing the relative path of the stage from the project root directory.
to the stage dir.
}
