% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\docType{class}
\name{dsoParams-class}
\alias{dsoParams-class}
\title{dsoParams: list with safe access}
\arguments{
\item{x}{empty, or a recursive list of lists which is converted to dsoParams}
}
\description{
A "dsoParams" and its constructor:
}
\section{Slots}{

\describe{
\item{\code{params}}{A list of lists containing parameters}
}}

\examples{
# initiating empty
params <- dsoParams()

# converting a list of list

params <- list()
params$a <- "bla"
params$b <- list()
params$b$c <- "blub"
params <- dsoParams(params)

dsoParams class

}
