% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{reload,dsoParams-method}
\alias{reload,dsoParams-method}
\title{reload dso params}
\usage{
\S4method{reload}{dsoParams}(object, env = caller_env())
}
\arguments{
\item{object}{dsoParams object}

\item{env}{environment in which object is located, caller_env() by default}
}
\description{
reloads the current dsoParams config into object
}
