% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dsoParams.R
\name{reload,dsoParams-method}
\alias{reload,dsoParams-method}
\title{reload dso params}
\usage{
\S4method{reload}{dsoParams}(object)
}
\arguments{
\item{object}{dsoParams object}
}
\description{
reloads the current dsoParams config into object
}
