% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{Reload dsoParams}
\usage{
reload(params, env = caller_env())
}
\arguments{
\item{params}{dsoParams object}

\item{env}{environment in which object is located, caller_env() by default}
}
\value{
The updated dsoParams object.
}
\description{
Reloads the current dsoParams configuration into the object.
}
